Integrate field binding regressions into existing tests
